// src/types/baserow.ts
export type Results = {
	Name: string;
	Location: string;
	Type: string;
	Rating: number;
	City: string;
	Country: string;
}


export type Headers = {
	id: number;
    table_id: number;
    name: string;
    order: number;
    type: string;
    primary: boolean;
    read_only: boolean;
    text_default: string;
}